app/services/task-service/requirements.txt
# ============================================================================
# Task Service - Python Dependencies
# ============================================================================
# Production dependencies for the task-service microservice
# Install with: pip install -r requirements.txt

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi==0.104.1                    # Modern, fast web framework with auto-docs
uvicorn[standard]==0.24.0           # ASGI server (runs FastAPI apps)
                                    # [standard] includes extra dependencies for performance

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy==2.0.23                  # SQL toolkit and ORM (Object-Relational Mapping)
psycopg2-binary==2.9.9              # PostgreSQL adapter for Python
alembic==1.12.1                     # Database migration tool (for schema changes)

# ============================================================================
# DATA VALIDATION & SETTINGS
# ============================================================================
pydantic==2.5.0                     # Data validation using Python type hints
pydantic-settings==2.1.0            # Settings management from environment variables

# ============================================================================
# OBSERVABILITY - OpenTelemetry (Distributed Tracing)
# ============================================================================
# OpenTelemetry = Industry standard for observability (your CV: comprehensive observability)
opentelemetry-api==1.21.0                          # Core OpenTelemetry API
opentelemetry-sdk==1.21.0                          # OpenTelemetry SDK
opentelemetry-instrumentation-fastapi==0.42b0      # Auto-instrument FastAPI
opentelemetry-instrumentation-sqlalchemy==0.42b0   # Auto-instrument SQLAlchemy
opentelemetry-instrumentation-requests==0.42b0     # Auto-instrument HTTP requests
opentelemetry-exporter-otlp==1.21.0                # Export traces to collectors (Jaeger, etc.)

# ============================================================================
# HTTP CLIENT (for making API calls)
# ============================================================================
httpx==0.25.1                       # Modern async HTTP client
requests==2.31.0                    # Traditional HTTP library (for compatibility)

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0                # Load environment variables from .env file
python-jose[cryptography]==3.3.0   # JWT token handling
passlib[bcrypt]==1.7.4              # Password hashing
python-multipart==0.0.6             # For form data parsing

# ============================================================================
# TESTING (Development Dependencies)
# ============================================================================
pytest==7.4.3                       # Testing framework
pytest-asyncio==0.21.1              # Async test support
pytest-cov==4.1.0                   # Code coverage reports
httpx==0.25.1                       # For testing HTTP endpoints

# ============================================================================
# CODE QUALITY (Development Dependencies)
# ============================================================================
black==23.11.0                      # Code formatter
flake8==6.1.0                       # Linting
mypy==1.7.0                         # Type checking
bandit==1.7.5                       # Security linting

# ============================================================================
# VERSION PINNING EXPLANATION
# ============================================================================
# We use exact versions (==) for:
# 1. Reproducibility - Same versions on all environments
# 2. Security - Know exactly what's running
# 3. Stability - Avoid breaking changes
#
# In your CV: "evergreen pipelines with Renovate/Dependabot"
# These tools will automatically update these versions safely!
